Comments Thomas 11/6/2013
Christine, Your approach is amazing. You combined your interest for the class with the need to immerse yourself in a new physical environment. And for that, you engage in the cyberworld. As a follow up other people might want to know how it has worked in the end and how the physical and the virtual worlds might have cross-nurtered ! Do you think you can integrate faster through online interaction? Why ?